+++ src/components/StyledProvider/index.js
@@ -3,10 +3,10 @@ import { ThemeProvider as StyledThemeProvider } from 'styled-components'
 import PropTypes from 'prop-types'
 import { createMuiTheme } from '@material-ui/core/styles'
 import { StylesProvider, ThemeProvider } from '@material-ui/styles'
-import theme from 'theme'
+import defaultTheme from 'theme'
 
-const StyledProvider = ({ children, otherTheme = {} }) => {
-  const mergedTheme = Object.assign(theme, otherTheme)
+const StyledProvider = ({ children, theme = {} }) => {
+  const mergedTheme = Object.assign(defaultTheme, theme)
   const finalTheme = createMuiTheme(mergedTheme)
 
   return (
@@ -23,7 +23,7 @@ StyledProvider.propTypes = {
    * Your custom theme. It has to respect Material-ui theming:
    * https://material-ui.com/customization
    */
-  otherTheme: PropTypes.object,
+  theme: PropTypes.object,
 }
 
 export default StyledProvider