+++ stories/utils/StyledProvider.js
@@ -1,0 +1,8 @@
+import { storiesOf } from '@storybook/react'
+
+import StyledProvider from 'components/StyledProvider'
+import createApiStory from '../helpers/createApiStory'
+
+const story = storiesOf('Utils|StyledProvider', module)
+
+createApiStory(story, StyledProvider)